+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Magic link verification: fix timezone-naive vs aware datetime comparison (SQLite stores naive)
+
 ## [1.5.0] - 2026-01-05
 
 ### Added